musicservice/requirements.txt
astroid==2.0.4
Django==2.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
six==1.11.0
wrapt==1.10.11
